Refactor S3 and POSIX datastores to share code
Shared code moved to FileLikeDatastore.
