+++ python/lsst/daf/butler/datastores/posixDatastore.py
@@ -28,33 +28,16 @@ import logging
 import os
 import shutil
 
-from collections import namedtuple
+from lsst.daf.butler import DatasetTypeNotSupportedError
 
-from lsst.daf.butler import (
-    Config,
-    Constraints,
-    DatabaseDict,
-    DatasetTypeNotSupportedError,
-    Datastore,
-    DatastoreConfig,
-    DatastoreValidationError,
-    FileDescriptor,
-    FileTemplates,
-    FileTemplateValidationError,
-    FormatterFactory,
-    LocationFactory,
-    StorageClassFactory,
-    StoredFileInfo,
-)
-
-from lsst.daf.butler.core.repoRelocation import replaceRoot
+from .fileLikeDatastore import FileLikeDatastore
 from lsst.daf.butler.core.safeFileIo import safeMakeDir
-from lsst.daf.butler.core.utils import transactional, getInstanceOf
+from lsst.daf.butler.core.utils import transactional
 
 log = logging.getLogger(__name__)
 
 
-class PosixDatastore(Datastore):
+class PosixDatastore(FileLikeDatastore):
     """Basic POSIX filesystem backed Datastore.
 
     Attributes
@@ -93,144 +76,13 @@ class PosixDatastore(Datastore):
     absolute path. Can be None if no defaults specified.
     """
 
-    RecordTuple = namedtuple("PosixDatastoreRecord", ["formatter", "path",
-                                                      "storage_class", "checksum",
-                                                      "file_size"])
-
-    @classmethod
-    def setConfigRoot(cls, root, config, full, overwrite=True):
-        """Set any filesystem-dependent config options for this Datastore to
-        be appropriate for a new empty repository with the given root.
-
-        Parameters
-        ----------
-        root : `str`
-            Filesystem path to the root of the data repository.
-        config : `Config`
-            A `Config` to update. Only the subset understood by
-            this component will be updated. Will not expand
-            defaults.
-        full : `Config`
-            A complete config with all defaults expanded that can be
-            converted to a `DatastoreConfig`. Read-only and will not be
-            modified by this method.
-            Repository-specific options that should not be obtained
-            from defaults when Butler instances are constructed
-            should be copied from ``full`` to ``config``.
-        overwrite : `bool`, optional
-            If `False`, do not modify a value in ``config`` if the value
-            already exists.  Default is always to overwrite with the provided
-            ``root``.
-
-        Notes
-        -----
-        If a keyword is explicitly defined in the supplied ``config`` it
-        will not be overridden by this method if ``overwrite`` is `False`.
-        This allows explicit values set in external configs to be retained.
-        """
-        Config.updateParameters(DatastoreConfig, config, full,
-                                toUpdate={"root": root},
-                                toCopy=("cls", ("records", "table")), overwrite=overwrite)
-
     def __init__(self, config, registry, butlerRoot=None):
-        super().__init__(config, registry)
-        if "root" not in self.config:
-            raise ValueError("No root directory specified in configuration")
-
-        # Name ourselves either using an explicit name or a name
-        # derived from the (unexpanded) root
-        if "name" in self.config:
-            self.name = self.config["name"]
-        else:
-            self.name = "POSIXDatastore@{}".format(self.config["root"])
-
-        # Support repository relocation in config
-        self.root = replaceRoot(self.config["root"], butlerRoot)
+        super().__init__(config, registry, butlerRoot)
 
         if not os.path.isdir(self.root):
             if "create" not in self.config or not self.config["create"]:
                 raise ValueError(f"No valid root at: {self.root}")
             safeMakeDir(self.root)
-
-        self.locationFactory = LocationFactory(self.root)
-        self.formatterFactory = FormatterFactory()
-        self.storageClassFactory = StorageClassFactory()
-
-        # Now associate formatters with storage classes
-        self.formatterFactory.registerFormatters(self.config["formatters"],
-                                                 universe=self.registry.dimensions)
-
-        # Read the file naming templates
-        self.templates = FileTemplates(self.config["templates"],
-                                       universe=self.registry.dimensions)
-
-        # And read the constraints list
-        constraintsConfig = self.config.get("constraints")
-        self.constraints = Constraints(constraintsConfig, universe=self.registry.dimensions)
-
-        # Storage of paths and formatters, keyed by dataset_id
-        types = {"path": str, "formatter": str, "storage_class": str,
-                 "file_size": int, "checksum": str, "dataset_id": int}
-        lengths = {"path": 256, "formatter": 128, "storage_class": 64,
-                   "checksum": 128}
-        self.records = DatabaseDict.fromConfig(self.config["records"], types=types,
-                                               value=self.RecordTuple, key="dataset_id",
-                                               lengths=lengths, registry=registry)
-
-    def __str__(self):
-        return self.root
-
-    def addStoredFileInfo(self, ref, info):
-        """Record internal storage information associated with this
-        `DatasetRef`
-
-        Parameters
-        ----------
-        ref : `DatasetRef`
-            The Dataset that has been stored.
-        info : `StoredFileInfo`
-            Metadata associated with the stored Dataset.
-        """
-        self.records[ref.id] = self.RecordTuple(formatter=info.formatter, path=info.path,
-                                                storage_class=info.storageClass.name,
-                                                checksum=info.checksum, file_size=info.size)
-
-    def removeStoredFileInfo(self, ref):
-        """Remove information about the file associated with this dataset.
-
-        Parameters
-        ----------
-        ref : `DatasetRef`
-            The Dataset that has been removed.
-        """
-        del self.records[ref.id]
-
-    def getStoredFileInfo(self, ref):
-        """Retrieve information associated with file stored in this
-        `Datastore`.
-
-        Parameters
-        ----------
-        ref : `DatasetRef`
-            The Dataset that is to be queried.
-
-        Returns
-        -------
-        info : `StoredFileInfo`
-            Stored information about this file and its formatter.
-
-        Raises
-        ------
-        KeyError
-            Dataset with that id can not be found.
-        """
-        record = self.records.get(ref.id, None)
-        if record is None:
-            raise KeyError("Unable to retrieve formatter associated with Dataset {}".format(ref.id))
-        # Convert name of StorageClass to instance
-        storageClass = self.storageClassFactory.getStorageClass(record.storage_class)
-        return StoredFileInfo(record.formatter, record.path, storageClass,
-                              checksum=record.checksum, size=record.file_size)
 
     def exists(self, ref):
         """Check if the dataset exists in the datastore.
@@ -245,14 +97,9 @@ class PosixDatastore(Datastore):
         exists : `bool`
             `True` if the entity exists in the `Datastore`.
         """
-        # Get the file information (this will fail if no file)
-        try:
-            storedFileInfo = self.getStoredFileInfo(ref)
-        except KeyError:
+        location, _ = self._get_dataset_location_info(ref)
+        if location is None:
             return False
-
-        # Use the path to determine the location
-        location = self.locationFactory.fromPath(storedFileInfo.path)
         return os.path.exists(location.path)
 
     def get(self, ref, parameters=None):
@@ -280,16 +127,9 @@ class PosixDatastore(Datastore):
         ValueError
             Formatter failed to process the dataset.
         """
-        log.debug("Retrieve %s from %s with parameters %s", ref, self.name, parameters)
-
-        # Get file metadata and internal metadata
-        try:
-            storedFileInfo = self.getStoredFileInfo(ref)
-        except KeyError as e:
-            raise FileNotFoundError("Could not retrieve Dataset {}".format(ref)) from e
-
-        # Use the path to determine the location
-        location = self.locationFactory.fromPath(storedFileInfo.path)
+        location, formatter, storedFileInfo, assemblerParams, \
+            component, readStorageClass = self._prepare_for_get(ref,
+                                                                parameters)
 
         # Too expensive to recalculate the checksum on fetch
         # but we can check size and existence
@@ -302,36 +142,12 @@ class PosixDatastore(Datastore):
             raise RuntimeError("Integrity failure in Datastore. Size of file {} ({}) does not"
                                " match recorded size of {}".format(location.path, size, storedFileInfo.size))
 
-        # We have a write storage class and a read storage class and they
-        # can be different for concrete composites.
-        readStorageClass = ref.datasetType.storageClass
-        writeStorageClass = storedFileInfo.storageClass
-
-        # Check that the supplied parameters are suitable for the type read
-        readStorageClass.validateParameters(parameters)
-
-        # Is this a component request?
-        component = ref.datasetType.component()
-
-        formatter = getInstanceOf(storedFileInfo.formatter,
-                                  FileDescriptor(location, readStorageClass=readStorageClass,
-                                                 storageClass=writeStorageClass, parameters=parameters))
-        formatterParams, assemblerParams = formatter.segregateParameters()
         try:
             result = formatter.read(component=component)
         except Exception as e:
             raise ValueError(f"Failure from formatter '{formatter.name()}' for Dataset {ref.id}") from e
 
-        # Process any left over parameters
-        if assemblerParams:
-            result = readStorageClass.assembler().handleParameters(result, assemblerParams)
-
-        # Validate the returned data type matches the expected data type
-        pytype = readStorageClass.pytype
-        if pytype and not isinstance(result, pytype):
-            raise TypeError("Got type {} from formatter but expected {}".format(type(result), pytype))
-
-        return result
+        return self._post_process_get(result, readStorageClass, assemblerParams)
 
     @transactional
     def put(self, inMemoryDataset, ref):
@@ -359,36 +175,7 @@ class PosixDatastore(Datastore):
         allow `ChainedDatastore` to put to multiple datastores without
         requiring that every datastore accepts the dataset.
         """
-        datasetType = ref.datasetType
-        storageClass = datasetType.storageClass
-
-        # Sanity check
-        if not isinstance(inMemoryDataset, storageClass.pytype):
-            raise TypeError("Inconsistency between supplied object ({}) "
-                            "and storage class type ({})".format(type(inMemoryDataset),
-                                                                 storageClass.pytype))
-
-        # Confirm that we can accept this dataset
-        if not self.constraints.isAcceptable(ref):
-            # Raise rather than use boolean return value.
-            raise DatasetTypeNotSupportedError(f"Dataset {ref} has been rejected by this datastore via"
-                                               " configuration.")
-
-        # Work out output file name
-        try:
-            template = self.templates.getTemplate(ref)
-        except KeyError as e:
-            raise DatasetTypeNotSupportedError(f"Unable to find template for {ref}") from e
-
-        location = self.locationFactory.fromPath(template.format(ref))
-
-        # Get the formatter based on the storage class
-        try:
-            formatter = self.formatterFactory.getFormatter(ref,
-                                                           FileDescriptor(location,
-                                                                          storageClass=storageClass))
-        except KeyError as e:
-            raise DatasetTypeNotSupportedError(f"Unable to find formatter for {ref}") from e
+        location, formatter = self._prepare_for_put(inMemoryDataset, ref)
 
         storageDir = os.path.dirname(location.path)
         if not os.path.isdir(storageDir):
@@ -508,66 +295,9 @@ class PosixDatastore(Datastore):
         checksum = self.computeChecksum(fullPath)
         stat = os.stat(fullPath)
         size = stat.st_size
-        self.registry.addDatasetLocation(ref, self.name)
 
-        # Associate this dataset with the formatter for later read.
-        fileInfo = StoredFileInfo(formatter, path, ref.datasetType.storageClass,
-                                  size=size, checksum=checksum)
-        # TODO: this is only transactional if the DatabaseDict uses
-        #       self.registry internally.  Probably need to add
-        #       transactions to DatabaseDict to do better than that.
-        self.addStoredFileInfo(ref, fileInfo)
-
-        # Register all components with same information
-        for compRef in ref.components.values():
-            self.registry.addDatasetLocation(compRef, self.name)
-            self.addStoredFileInfo(compRef, fileInfo)
-
-    def getUri(self, ref, predict=False):
-        """URI to the Dataset.
-
-        Parameters
-        ----------
-        ref : `DatasetRef`
-            Reference to the required Dataset.
-        predict : `bool`
-            If `True`, allow URIs to be returned of datasets that have not
-            been written.
-
-        Returns
-        -------
-        uri : `str`
-            URI string pointing to the Dataset within the datastore. If the
-            Dataset does not exist in the datastore, and if ``predict`` is
-            `True`, the URI will be a prediction and will include a URI
-            fragment "#predicted".
-            If the datastore does not have entities that relate well
-            to the concept of a URI the returned URI string will be
-            descriptive. The returned URI is not guaranteed to be obtainable.
-
-        Raises
-        ------
-        FileNotFoundError
-            A URI has been requested for a dataset that does not exist and
-            guessing is not allowed.
-
-        """
-        # if this has never been written then we have to guess
-        if not self.exists(ref):
-            if not predict:
-                raise FileNotFoundError("Dataset {} not in this datastore".format(ref))
-
-            template = self.templates.getTemplate(ref)
-            location = self.locationFactory.fromPath(template.format(ref) + "#predicted")
-        else:
-            # If this is a ref that we have written we can get the path.
-            # Get file metadata and internal metadata
-            storedFileInfo = self.getStoredFileInfo(ref)
-
-            # Use the path to determine the location
-            location = self.locationFactory.fromPath(storedFileInfo.path)
-
-        return location.uri
+        # Update the registry
+        self._register_dataset(ref, formatter, path, size, checksum)
 
     def remove(self, ref):
         """Indicate to the Datastore that a Dataset can be removed.
@@ -590,101 +320,16 @@ class PosixDatastore(Datastore):
             Attempt to remove a dataset that does not exist.
         """
         # Get file metadata and internal metadata
+        location, storefFileInfo = self._get_dataset_location_info(ref)
+        if location is None:
+            raise FileNotFoundError(f"Requested dataset ({ref}) does not exist")
 
-        try:
-            storedFileInfo = self.getStoredFileInfo(ref)
-        except KeyError as e:
-            raise FileNotFoundError("Requested dataset ({}) does not exist".format(ref)) from e
-        location = self.locationFactory.fromPath(storedFileInfo.path)
         if not os.path.exists(location.path):
-            raise FileNotFoundError("No such file: {0}".format(location.uri))
+            raise FileNotFoundError(f"No such file: {location.uri}")
         os.remove(location.path)
 
         # Remove rows from registries
-        self.removeStoredFileInfo(ref)
-        self.registry.removeDatasetLocation(self.name, ref)
-        for compRef in ref.components.values():
-            self.registry.removeDatasetLocation(self.name, compRef)
-            self.removeStoredFileInfo(compRef)
-
-    def transfer(self, inputDatastore, ref):
-        """Retrieve a Dataset from an input `Datastore`,
-        and store the result in this `Datastore`.
-
-        Parameters
-        ----------
-        inputDatastore : `Datastore`
-            The external `Datastore` from which to retreive the Dataset.
-        ref : `DatasetRef`
-            Reference to the required Dataset in the input data store.
-
-        """
-        assert inputDatastore is not self  # unless we want it for renames?
-        inMemoryDataset = inputDatastore.get(ref)
-        return self.put(inMemoryDataset, ref)
-
-    def validateConfiguration(self, entities, logFailures=False):
-        """Validate some of the configuration for this datastore.
-
-        Parameters
-        ----------
-        entities : iterable of `DatasetRef`, `DatasetType`, or `StorageClass`
-            Entities to test against this configuration.  Can be differing
-            types.
-        logFailures : `bool`, optional
-            If `True`, output a log message for every validation error
-            detected.
-
-        Raises
-        ------
-        DatastoreValidationError
-            Raised if there is a validation problem with a configuration.
-            All the problems are reported in a single exception.
-
-        Notes
-        -----
-        This method checks that all the supplied entities have valid file
-        templates and also have formatters defined.
-        """
-
-        templateFailed = None
-        try:
-            self.templates.validateTemplates(entities, logFailures=logFailures)
-        except FileTemplateValidationError as e:
-            templateFailed = str(e)
-
-        formatterFailed = []
-        for entity in entities:
-            try:
-                self.formatterFactory.getFormatterClass(entity)
-            except KeyError as e:
-                formatterFailed.append(str(e))
-                if logFailures:
-                    log.fatal("Formatter failure: %s", e)
-
-        if templateFailed or formatterFailed:
-            messages = []
-            if templateFailed:
-                messages.append(templateFailed)
-            if formatterFailed:
-                messages.append(",".join(formatterFailed))
-            msg = ";\n".join(messages)
-            raise DatastoreValidationError(msg)
-
-    def getLookupKeys(self):
-        # Docstring is inherited from base class
-        return self.templates.getLookupKeys() | self.formatterFactory.getLookupKeys() | \
-            self.constraints.getLookupKeys()
-
-    def validateKey(self, lookupKey, entity):
-        # Docstring is inherited from base class
-        # The key can be valid in either formatters or templates so we can
-        # only check the template if it exists
-        if lookupKey in self.templates:
-            try:
-                self.templates[lookupKey].validateTemplate(entity)
-            except FileTemplateValidationError as e:
-                raise DatastoreValidationError(e) from e
+        self._remove_from_registry(ref)
 
     @staticmethod
     def computeChecksum(filename, algorithm="blake2b", block_size=8192):
